Update core.py
keras.callbacks.TensorBoard expects a numpy item:

/usr/local/lib/python3.6/dist-packages/keras/callbacks.py in on_epoch_end(self, epoch, logs)
    860             summary = tf.Summary()
    861             summary_value = summary.value.add()
--> 862             summary_value.simple_value = value.item()
    863             summary_value.tag = name
    864             self.writer.add_summary(summary, epoch)

AttributeError: 'int' object has no attribute 'item'

+++ rl/core.py
@@ -208,9 +208,9 @@ class Agent(object):
 
                     # This episode is finished, report and reset.
                     episode_logs = {
-                        'episode_reward': episode_reward,
-                        'nb_episode_steps': episode_step,
-                        'nb_steps': self.step,
+                        'episode_reward': np.float64(episode_reward),
+                        'nb_episode_steps': np.uint32(episode_step),
+                        'nb_steps': np.uint32(self.step),
                     }
                     callbacks.on_episode_end(episode, episode_logs)
